refactor scripts into modules
