+++ letterbot/connection.py
